% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euro_area_indicators.R
\docType{data}
\name{euro_area_indicators}
\alias{euro_area_indicators}
\title{Six Euro Area Monetary Indicators}
\format{A matrix containing as columns six quarterly time series
	ranging from Q1,1999 to Q4,2017.}
\source{
ECB Real Time DataBase 'https://sdw.ecb.europa.eu/browse.do?node=9689716'.
}
\usage{
euro_area_indicators
}
\description{
This data set gives thre quarterly time series of 
  real gross domestic product, M3 aggregate, M1 aggregate, 
  inflation rate (HICP), unemployment rate and long-term interest rate
  for the Euro Area from Q1,1999 to Q4,2017, according to the ECB Real Time
  DataBase (RTDB).
}
\details{
Documentation of the dataset 'euro_area_indicators'
}
\references{
Farne', M., Montanari, A., 2018.
	A bootstrap test to detect prominent Granger-causalities across frequencies.
	<arXiv:1803.00374>, \emph{Submitted}.

Euro Area Real Time Database documentation.
	'http://sdw.ecb.europa.eu/web/docu/rtdb_docu.pdf'
}
\keyword{datasets}
